+++ webtest/resources.go
@@ -3,12 +3,10 @@ package main
 import (
 	"fmt"
 	"io"
-	"io/ioutil"
-	"log"
 	"net/http"
-	"os"
 )
 
+// Resouse pass
 const (
 	HELLO_WORLD_PAGE           = "../html/index.html"
 	FORBIDDEN_PAGE             = "../html/forbidden.html"
@@ -19,12 +17,7 @@ const (
 	UNKNOWN_ERROR_PAGE         = "../html/unknown_error.html"
 )
 
-const (
-	RED   = "\033[31m"
-	GREEN = "\033[32m"
-	RESET = "\033[0m"
-)
-
+// Field name
 const (
 	HOST          = "Host"
 	USERAGENT     = "User-Agent"
@@ -34,12 +27,6 @@ const (
 	CONNECTION    = "Connection"
 	AUTHORIZATION = "Authorization"
 )
-
-// http.Clientの説明にグローバルで使用すべきと書いてあった(詳しくは分からん)
-// 毎度作り直すことによる弊害の推測:
-//   作ることのオーバヘッド
-//   接続のキャッシュ情報が捨てられることによるリーク
-var client = &http.Client{}
 
 const (
 	PreURI = "http://localhost:"
@@ -75,44 +62,4 @@ func (t testCase) NewRequest() (*http.Request, error) {
 		req.Header.Add(key, value)
 	}
 	return req, nil
-}
-
-//// Request: 引数で渡された情報を元にリクエストを作成します.
-//func Request(method, port, url string, addQuery, addFields map[string]string, body io.Reader) *http.Response {
-//	req, err := http.NewRequest(method, PreURI+port+url, body)
-//	if err != nil {
-//		log.Fatalf("fail to send request: %v", err)
-//	}
-
-//	// URLパラメータ追加
-//	q := req.URL.Query()
-//	for key, value := range addQuery {
-//		q.Add(key, value)
-//	}
-//	req.URL.RawQuery = q.Encode()
-
-//	// ヘッダーフィールド追加
-//	for key, value := range addFields {
-//		req.Header.Add(key, value)
-//	}
-
-//	resp, err := client.Do(req)
-//	if err != nil {
-//		log.Fatalf("fail to get response: %v", err)
-//	}
-//	return resp
-//}
-
-// FileContents: fileNameで指定されたパスのファイルの中身を[]byteに詰めて返します.
-func FileContents(fileName string) []byte {
-	file, err := os.Open(fileName)
-	if err != nil {
-		log.Fatalf("FileBytes: %v", err)
-	}
-	defer file.Close()
-	srcBytes, err := ioutil.ReadAll(file)
-	if err != nil {
-		log.Fatalf("FileBytes: %v", err)
-	}
-	return srcBytes
 }
