+++ webtest/test_GET.go
@@ -1,10 +1,7 @@
 package main
 
 import (
-	"bytes"
 	"fmt"
-	"io"
-	"log"
 	"net/http"
 )
 
@@ -93,43 +90,6 @@ func testGET() {
 			wantStatusCode: tt.wantStatusCode,
 			wantBody:       tt.wantBody,
 		}
-		test(tc)
-	}
-}
-
-func test(tc testCase) {
-	fmt.Print("[ " + tc.name + " ] ")
-	var wasErr bool
-
-	req, err := tc.NewRequest()
-	if err != nil {
-		log.Fatalf("failt to create request: %v", err)
-	}
-
-	// リクエスト and レスポンス受け取り
-	resp, err := client.Do(req)
-	if err != nil {
-		log.Fatalf("fail to send request: %v", err)
-	}
-	defer resp.Body.Close()
-
-	// ステータスコードの確認
-	if resp.StatusCode != tc.wantStatusCode {
-		wasErr = true
-		fmt.Printf("actual_status: %v, expect_status: %v\n", resp.StatusCode, tc.wantStatusCode)
-	}
-
-	// ボディの確認
-	resposenseBody, err := io.ReadAll(resp.Body)
-	if err != nil {
-		log.Fatalf("fail to get read body: %v", err)
-	}
-	if bytes.Compare(resposenseBody, tc.wantBody) != 0 {
-		wasErr = true
-		fmt.Printf("actual_body: %v, expect_body: %v\n", resp.StatusCode, tc.wantStatusCode)
-	}
-
-	if !wasErr {
-		fmt.Println(GREEN, "ok", RESET)
+		Do(tc)
 	}
 }
